Add files via upload
The first page in the project, a Python file,
The code contains a category of the courses that I am going to present on the site and the users can learn the following courses. Each course includes the name of the course and a picture to complement it and a short description of the course.
The second category is news or articles, which will present important articles and news in the world of computer science and the user will always be updated. The article or news includes a title and is related to whom and on what topic
LOCAL HOST )127.0.0.1 is used to see the updates I make to the project at each stage.